gsheet/skills/scripts/pyproject.toml
[project]
name = "gsheet-skill"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
]
